Ignore strides of size-1 dimensions when determining if tensor is contiguous
If a dimension has a size of 1, only index 0 is valid and since `0 * stride = 0`
for all strides, we can ignore the stride when determining whether a tensor is
contiguous.

This allows more code to use fast paths for contiguous tensors. In particular
when new dimensions are inserted in the middle of a tensor by `Unsqueeze`
operations in a model, they no longer become non-contiguous.