Support using matmul for Einsum equations with 3 or more inputs
Previously Einsum equations with more than two inputs were handled via a
fallback to a naive implementation that was much less efficient than a matmul if
a reduction was required. In this commit, equations are split into a path where
each step handles one or two inputs. Each step will use matmul, multiply or
reduce-sum as appropriate.

The generated path always processes inputs in left-to-right order. This can be
much less efficient than the optimal contraction order. Using a better algorithm
is left as a future exercise.