Implement `If` operator
Add infrastructure to support control flow operators which execute subgraphs and
use it to implement the `If` operator.

The implementation of control flow operators follows ONNX by supporting
graphs as operator attributes. These graphs may refer to (_capture_)
values from the parent graph at runtime, like a closure. The `If`
operator has two subgraphs which take no inputs but instead use captures
to access values from the parent graph.

Operator inputs in RTen are represented as a list of graph-local node IDs rather
than names. Therefore an operator's input list cannot directly refer to captured
values by name. Instead when an operator input is encountered that is a capture,
a value node is created in the subgraph to represent it and the ID of that value
node is added to a list of `captures` for the graph.

At runtime when resolving the value of a node it is first looked up by ID
in inputs and locals for the current graph and if that fails, a lookup by name
is performed in the capture environment.

 - Add `If` operator to RTen model schema

 - Support deserializing operators with attribute values that are graphs

 - Extend `Operator` trait with a `run_subgraph` method that is like `run` but
   takes a capture environment as an extra argument and returns a `RunError`
   instead of an `OpError`.

 - Implement `If` operator that uses `run_subgraph`

 - Support running operators with subgraphs in `Graph::run_plan`