Allow BPE tokens containing unused tokens in vocab
Support loading tokenizers which contain entries in the `vocab` map that do not
appear in either `merges` or `added_tokens`. The TrOCR base model on Hugging
Face (https://huggingface.co/microsoft/trocr-base-printed) has an
`<|endoftext|>` token in the vocab which does not appear in the `merges` or
`added_tokens` fields.