Fix `TensorDataBuilder` returning wrong offset
When adding a new tensor, the returned offset was the end of the previous tensor
instead of the start of the new tensor. These offsets may differ if padding
bytes are added for alignment. The problem went unnoticed before because all the
supported data types had the same size (32bit), and that has now changed.