Move `normalize_image` to rten-imageproc and make it customizable
Several examples need to normalize an image with a mean and standard deviation
different from the imagenet values. Add a function in rten-imageproc with this
flexibility. The function has been added to the rten-imageproc crate because
it doesn't have any dependencies on specific image formats.

In the process the implementation was also revised to make it more efficient.