Make GEMM outer loops generic over input types
This was straightforward except for the thread-local packing buffers. These use
`static` which cannot be generic. Since all the element types that we need to
support for GEMM operations are simple `Copy` types for which any bit pattern is
valid, the solution is to pick such a type with a large-enough alignment and
transmute the packing buffers' contents.