Replace old tensor slicing methods with new API
Remove the old `TensorBase::{slice, slice_mut, slice_dyn, slice_mut_dyn}`
implementations and replace them with the `slice_with` APIs that infer the
result layout type based on the input layout and slice range.

This now leaves the API with a smaller and easier to use set of methods for
slicing, which automatically infer the output view's layout.