Add custom `Debug` impl for `TensorBase`
Add a `Debug` impl for TensorBase that outputs the data in logical order, elided
if sufficiently long, followed by shape and stride info.

The main `tensor.rs` module is growing very long (~3.8K lines, of which 40% are
tests), so start the process of moving some impls and their tests into separate
modules.