Convert `NodeId` from an alias for `usize` to u32-sized opaque type
Align the representation of NodeId values with the FlatBuffers schema, which
uses `uint` (u32) as the type, except for operator inputs and outputs where
`int`s are used, with negative values representing missing optional inputs.

Since node IDs are in the range `[0, i32::MAX]` we can reserve higher values as
a niche, making `Option<NodeId>` the same size as a `NodeId`.