Support passing captures by-value to subgraphs
When preparing the capture environment to pass to an operator with subgraphs,
if a captured value is not going to be needed by subsequent operators in the
current graph then it can be passed by value rather than by reference. In
the subgraph, this allows the value to be (potentially) used as an in-place
input.

 - Pass `CaptureEnv` to `Graph::run_subgraph` by value rather than by reference

 - Add map of by-value captures to `CaptureEnv`

 - Add `CaptureEnv::child` method which creates a child environment with no
   captures of its own. This is not currently used but will be needed by loop
   operators which run a subgraph more than once and need to obtain a fresh
   capture environment for each iteration.

Part of https://github.com/robertknight/rten/issues/399.