Reuse existing output node when replacing operator with a fusion
When replacing a subgraph with a fused operator, reuse the output value node
from the subgraph instead of creating a new one. This preserves metadata such as
the name and shape associated with that value node. Also it simplifies the code
by removing the need to replace all references to the previous output node with
the new one.

This improves the runtime of the Whisper example using the whisper-base model by
~25% by fixing an issue where fused Transpose + MatMul operations did not get
used.