Make `Graph::add_node` private again
Exposing `add_node` as a public method creates a hazard as it allows callers to
add nodes to the graph without post-processing steps that other methods do (eg.
`add_op`). Make this method private again and add a more specific
`add_constant_node` alternative to handle the single use case for it outside the
graph module.