Re-order execution plan to enable more operations to run in-place
After creating the initial execution plan by traversing the graph backwards from
the requested outputs, re-order it by traversing the plan forwards and choosing
a preferred operator to run from the frontier of runnable operators at each
step. The initial logic for picking the next operator is to choose the first
non-in-place operator if there is one (eg. Shape) or the first operator
otherwise.

This re-ordering increases the likelihood that an operator which can run
in-place is able to do so, because its in-place input is less likely to be
required by a non in-place op that runs later.

Fixes https://github.com/robertknight/rten/issues/98
See also https://github.com/robertknight/rten/issues/399#issuecomment-2466625999.