Reserve KV cache capacity after the first model run
Hugging Face models with separate branches for the first and subsequent
iterations do not use the input KV cache buffer on the first run. Thus they did
not benefit from the pre-allocated capacity and ended up re-allocating a new KV
cache buffer on each run.

To resolve this, change the KV cache growth strategy to grow the buffer after
the model runs, if the capacity limit has been reached. Also replace the
hard-coded capacity with a growth strategy that doubles the capacity each time.
This amortizes the costs of copying the old KV cache into the new buffer.