Test running all operators in-place
In the existing test that runs all operators, run the model twice. The
first time the operator's main input is passed as a view, the second
time as an owned tensor. This enables running the operator in-place if
supported. This is an easy way to gain some additional coverage of
`Operator::run_in_place` implementations.