Convert `Normalizer` into a trait
Add a `Normalizer` trait which provides the common interface for normalizers and
rename the previous `Normalizer` struct to `BertNormalizer`, to align with the
name in tokenizer.json files. Change other parts of the tokenization process
to use the normalizer via a trait object.

Part of https://github.com/robertknight/rten/issues/427