Simplify fallback for reducing 2+ axes in a non-contiguous tensor
Shuffle the axes so the N reduced dimensions are last, then iterate over slices
of the inner N dims.

The implementation of `TensorBase::inner_iter_dyn` is essentially the same as
the previous code, but this could be improved by taking advantage of the fact
that only the start offset in the data changes on each iteration.