Unify kernels of normalization functions
The kernels of BatchNormalization, InstanceNormalization and LayerNormalization
are very similar. The operators differ in:

 - Which slices of the input are normalized (batch item, channel etc.)
 - Whether the mean and variance computed at inference time or estimates are
   pre-computed at training time
 - The granularity of the bias and scale factors (global, per channel,
   per element etc.)

These can be captured in a single kernel which computes:

```
Y[i] = (X[i] - mean) * scale * element_scale[i] + bias + element_bias[i]
```

On some slice of the input.

For the LayerNormalization operator this change has the benefit of removing a
pass over the data that subtracts the mean. Instead the mean is subtracted both
in the pass that computes the variance and the one that normalizes the output.