Make `TensorBase::reshaped` return maybe-owned view
Previously `reshaped` had a footgun where it would panic if the input view was
not contiguous. Make it copy the input instead in this case. The downside is
that consumers cannot chain `reshaped` into the middle of a sequence of view
operations because the lifetime of the result is now shorter.