Make `reshaped_mut` return error if view is not contiguous
Unlike `TensorBase::reshaped` we can't fall back to copying if the view is not
contiguous because this method is often used to obtain a mutable slice into a
larger tensor to write to, and it would be incorrect to write into a copy.

Instead make this method return a result so that callers are made aware they
need to consider this possibility.