Remove `pad` argument from `Simd::load_partial`, `simd_fold`, `simd_map`
Simplify usage of these functions by removing the `pad` argument and always
setting unused lanes to zero. This was already the case in most places, with a
few exceptions for reductions where using zero as a padding value affected the
final result. Handle this instead by using the accumulator for the
one-before-last update for the lanes that are unused in the final update.

As well as simplifying usage, this also allows for translating partial loads
into masked loads.