Better document in GEMM impl when output is potentially uninitialized
If `beta` is zero then the output is potentially uninitialized. This was
represented in the public API (`GemmExecutor::gemm_uninit`) but not internally
as the output was always transmuted to `&mut [OutT]` before calling `gemm_impl`.
Change the internals to use `MaybeUninit<OutT>` everywhere that the output is
potentially uninitialized.

This will help to avoid mistakes like the one fixed in
143f0a31dba8da2c20468f41eab4841180bd6fad.