Refactor rten-vecmath APIs to reduce boilerplate
Previously each operation in rten-vecmath was exposed as three functions: one
for scalars, one that read/wrote to different slices and one that mutated a
slice in place. This led to quite a bit of boilerplate for each function added.

Refactor this so that for unary functions rten-vecmath just exports a struct
which implements the operation. Dispatch is then handled by the caller using
trait methods of `SimdUnaryOp`.

As part of this, drop the `Simd*` prefix from type names and `vec_*` prefix from
functions. They were not informative since most functions/types in the crate had
them. The caller can rename the imports if they conflict.