Push handling of non-full width tiles down into GEMM kernels
For tiles with fewer than NR columns, the GEMM impl creates a temporary tile on
the stack, runs the kernel on that and copies the result into the real output
tile afterwards. For architectures with masked loads and stores more efficient
approaches are possible. As a step to enable this, move handling of non-full
width tiles into the arch-specific kernels.