Inline `simd_{op}` functions into `SimdUnaryOp::eval` impls
Previously the implementation of vectorized ops was split into two parts: a
`simd_{op}` internal function and the `SimdUnaryOp::eval` impl which provided a
way for SIMD dispatch to invoke `simd_{op}`. The main reason for the separation
was so that one SIMD operation could invoke another easily by calling eg.
`simd_exp(x)`.

Remove this separation by inlining the `simd_{op}` function into the `eval`
impl. To assist with one SIMD op calling another, add a `SimdUnaryOp::apply`
helper. This means that eg. softmax can calling exp using `Exp::apply(x)`.

An upside of this change is that consumers of the package can more easily call
the operations in their own vectorized ops. The downside is that it adds some
indirection for the reader.