Divide prepacked matmul inputs into depth blocks
Divide prepacked into depth blocks with a size that matches the depth block size
(`kc`) used during computation.

This allows for the generic GEMM code to be agnostic of the layout of panels
within a block, as it no longer needs to be able to slice panels along the depth
dimension. Instead it just uses a depth block index to pick a panel with the
pre-determined block size. This in turn gives the GEMM kernel freedom to choose
the layout within each panel.

This partly undoes https://github.com/robertknight/rten/pull/482.