Switch to row major layout for packed blocks of LHS / A
This change means that both packed and unpacked LHS inputs handled by the f32
GEMM kernel have unit column stride, which simplifies indexing calculations
when broadcasting elements of A in the main loop. As a result a bunch of `ADD`
instructions are eliminated from the hottest loop in the code.

The previous packing code has been kept around to enable comparisons of the
earlier approach to packing of the LHS (always pack, column major order) and
the new approach (only pack if not unit column stride, row major order). The
earlier approach follows the BLIS framework. The newer approach matches what
ORT and XNNPACK do.