Unify gemm APIs with and without bias arguments
Combine the `GemmExecutor` methods that take biases and those that don't. This
streamlines the API a little. When new GEMM "extensions" are added in future,
they should probably be part of an options struct to avoid having methods for
the cross-product of GEMM extensions.

This highlights a few existing places where code calls `GemmExecutor::gemm`
without a bias and then immediately adds one. This could likely be fused.