Print only one message per dynamic dim name without a specified size
Some models, such as those with KV-cache inputs, have a lot of inputs with the
same dynamic dim names. To make the CLI output less verbose, only print one
message per dimension name for which no size was specified.

A downside of this change is that if the user specifies a dim size with both an
input name and a dim name, and there is a match for the dim name but not the
input name, the output will not indicate which input didn't match. In most cases
though dim names are specified without input names.