Support quantization params in GEMM kernels, add basic i8 x u8 -> i32 kernel
Add arguments to `Kernel` trait methods to support passing zero point to
packing and kernel methods. Use this to implement a generic (and slow)
i8 x u8 -> i32 kernel. Convert the MatMulInteger operator to use
`GemmExecutor` for performing matmuls using this new kernel.

Several other parts of the code needed to have generic arguments added to
disambiguate which kind of `GemmExecutor` to instantiate.