Make GEMM functions return errors on invalid inputs
As the interface evolves the number of ways that inputs can be invalid is
growing. Change the interface to return an error instead of panicking to make
testing these cases easier.

Asserts around the size of the input were removed when validating pre-packed
inputs. The size is part of the pre-packed input, so a mismatch is not possible.