Add fallback for older Arm CPUs that don't support UDOT
This uses a sequence of instructions to emulate UDOT using widening multiplies
and pairwise adds.

In the process the helpers for instantiating `GemmExecutor` with different
kernels were refactored into a generic trait to make it easier to exercise each
of the available kernels for a given set of input and output types.