Refactor depthwise convolution to support arch/data type-specific kernels
Following the pattern of `GemmExecutor` and its kernels, create a
`DepthwiseConvExecutor` struct which abstracts the details of handling depthwise
convolutions for different architectures and data types. The top-level
convolution code then only needs a `DepthwiseConvExecutor<X, W, Y>: Default`
bound to ensure than an executor is available for the combination of input and
output types, instead of needing to have trait bounds that reflect the
implementation details.