Update range-limiting behavior in GEMM tests to match ORT quantizer
The ONNX quantization tools work around the saturation issue in the x64
`vpmaddubsw` instruction by limiting the range of the weights when
`reduce_range=True` is set. The GEMM tests however limited the range of the LHS
instead. This made the GEMM inputs generated different than what was encountered
when running an actual model. As a result it was possible to write kernels which
encountered saturation with real models quantized with `reduce_range=True`, but
still passed the tests.

Change the range-limiting in the tests to mirror what happens in actual models.