Implement ConvInteger operator
Implement the ConvInteger operator for all combinations of int8 input
signed-ness, with the caveat that combinations other than (input=int8,
weights=uint8) will result in a shift-conversion of the input and/or weights to
`input=int8`, `weights=uint8` which maps to the currently supported GEMM
kernels.

Depthwise convolution for int8 uses a naive conversion of the f32 depthwise
kernel and both have significant room for optimization.