Adjust zero point when converting batched matmul to non-batched
When a batched matmul `[A, M, K] x [K, N]` is converted to a non-batched matmul
with LHS shape `[A * M, K]` the zero point needs to be broadcast to match the
new row count.

This fixes an error when running the Segment Anything demo with a quantized
image encoder.