Avoid underflow when converting weights from i8 -> u8 in ConvInteger
In case the weights were not quantized using `reduce_range`, values might be <
-64. In that case we need to avoid underflow while minimizing the impact of
saturation that may occur.

The downside here is that we are adding an extra pass over the data on each
call. This could be fixed in future by adding more GEMM kernels to avoid
the need for the conversion, or doing the conversion as part of model load.

Part of https://github.com/robertknight/rten/issues/574.