Bump up x64 int8 tile size
Make `NR` 2x the vector width instead of 1x. This follows a similar change in
the Arm and wasm32 kernels. The tile sizes for the int8 -> int32 kernels now
align with the f32 kernels.

For AVX2 this improves performance for M=N=K=1024 by ~10%. For AVX512 with VNNI
the improvement is smaller (4% on a client Ice Lake i5).