Refactor and export `AssumeInit` trait
Move the `AssumeInit` trait used in lazy initialization of collections into its
own submodule, expand the documentation and export it for use in downstream
crates.

In the process the implementation for `Vec<MaybeUninit<T>>` was reworked to be
more sound. The previous transmute assumed that `Vec<MaybeUninit<T>>` and
`Vec<T>` have the same representation. I believe that is true in practice but
not guaranteed. The new implementation de-constructs and then re-constructs the
`Vec` in way that I believe should be sound.