Replace PtrLen/MutPtrLen with SrcDest for vectorized op inputs/outputs
Vectorized operations can operate either in-place on a single mutable slice, or
a pair of source and destination slices where the destination is uninitialized.
To handle both cases with the same implementation, slice-like types without the
aliasing guarantees were used. Replace this with a `SrcDest` type which models
the possible cases better in the type system and is a step towards reducing
unsafety in vectorized ops.

Part of https://github.com/robertknight/rten/issues/549.