Make SIMD vector length a runtime rather than compile-time constant
Make `Simd::LEN` an `Option<usize>` to represent the possibility that the vector
length is unknown at compile time, but a fixed size at runtime, and add a
`Simd::len` method which returns either the compile or runtime-constant size.

Runtime-determined SIMD vector types (eg. Arm SVE, RVV) are not currently
supported in Rust, but the expectation is that they will be eventually [^1].
This change is made in the hope that it will make it easier to support such
types in future.

[^1]: https://github.com/davidtwco/rfcs/blob/sized-hierarchy/text/3729-sized-hierarchy.md