Use stronger typing for `SimdUnaryOp` argument
Revise signature of `SimdUnaryOp::eval` so that it specifies the element type in
the SIMD vector. Casting between the generic type and the concrete type used by
the ISA is still required, but is now safer as the cast prevents accidental
changing of the element type.

A bonus is that it now becomes slightly easier to implement an operation for
different element types via a macro, as only the element type needs to varied
for each impl in many cases.