Support u16 vectors in new SIMD API
For x64 there aren't specific intrinsics for operations on u16 values in many
cases, but we rely on the fact that the i16 operations work. This is either
because the operation only cares about the lane width (splat, select etc.) or
that it works the same for signed/unsigned in two's complement (add, sub, low
half of multiplication).