Use separate lifetimes for source and destination of vectorized ops
Use separate lifetimes for the source and destination buffers of vectorized ops,
so that the compiler can understand that the source buffer is no borrowed after
an operation is dispatched.

This change is an enabler for creating a safe API to update the length of a
`Vec` after writing to the uninitialized portion.