Refactor int8 GEMM kernels to use dot product instructions via a trait
Refactor the int8 GEMM kernels to use a trait to access dot product
instructions. This trait is then implemented for witness types which ensure the
availability of the instructions used, following the approach of the new SIMD
API. This allows the `dot_product` method to be safe.

In the process the safety comments were updated for the SIMD GEMM / GEMV
kernels to reflect the new responsibilities for the caller.