Add `ExtendInit` utility to rten-vecmath
Non in-place operations in this crate take an uninitialized output slice, so the
caller has control over the output allocation and doesn't have to redundantly
initialize it. This has the downside that unsafe code has to be used to update
the length afterwards. Add an `ExtendInit` utility trait which handles this for
the common case of filling a `Vec`.