Parametrize `NumOps` by element type instead of SIMD type
Make the the generic argument to `NumOps` independent of the SIMD ISA. For
example `<Avx2Isa as Isa>::f32` now returns an `impl NumOps<f32>` instead of
`impl NumOps<float32x4_t`. This in turn will make it possible to add generic
methods to get the `NumOps` impl for a given element type, enabling writing
`SimdOp`s that are generic over the element type.