Make `simd_int8_gemv` a safe function
Make the entry point to the GEMV implementation a safe function, which checks
its preconditions and then dispatches to the appropriate (unsafe)
implementation.

This is now possible since the only remaining preconditions not enforced by the
type system are that the input and output dimensions are compatible, and that
the `a` matrix has the expected alignment.