Add `IntOps` trait
Add a trait for operations on SIMD vectors containing integers, whether signed
or unsigned. Move `SignedIntOps::shift_left` to the new trait and make
`SignedIntOps` a sub-trait of `IntOps`.

This creates a place to support operations like shifts, saturating addition etc.
for unsigned types in the future.

The `Isa::u{N}` methods currently still return an `impl NumOps`, but the plan is
for them to return `impl IntOps` in future.