Fix `pytorch_half_pixel` resize mode in Resize op
The formula was copied from the specification, but neglected to account for the
fact that `scale` in the Rust code is the scale going from output to input,
whereas in the spec it is the inverse scale from input to output. In the Rust
code the inverse scale is used in order to replace division with multiplication.