Add `TensorBase::slice_axis` and expose `slice_axis_mut`
These methods are useful for splitting a tensor into chunks along a given axis.

In the process `Layout::slice_axis` was changed to add bounds checks on the axis
and range, returning a `SliceError` if these are invalid. This in turn makes
`slice_axis` fail earlier and with a better error if the arguments are invalid.