Optimize BatchNormalization by avoiding tensor slicing
Since the tensor is contiguous and we only need the data for each chunk, we can
replace N * C `slice_mut` calls with a much cheaper `chunks_mut` iterator.

This made BatchNormalization ~10% faster on a MobileNet v4 model where there are
layers with many channels but a relatively small number of elements per channel.