Change Arm GEMM tile size to MR=4, NR=16
Change f32 GEMM tile size on Arm from 8x8 to 4x16. Practical tests on a handful
of devices (Apple M3, Raspberry Pi 5) found this to be faster and llvm-mca
analysis of the main loop reported that this computes the same number of output
elements in fewer cycles.

The speedup varies by CPU. On an M3 this makes f32 MatMul ops about 8% faster.