Fix "instruction requires: dotprod" error in Arm Linux build
On macOS using `udot` in inline assembly works when the calling function doesn't
have the `dotprod` target feature enabled. In a Linux arm64 build this is not
the case, resulting in an error:

```
error: instruction requires: dotprod
   --> src/gemm/kernels/aarch64.rs:462:18
    |
462 |                 "udot {result:v}.4s, {a:v}.16b, {b:v}.16b",
    |                  ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
    |
```

Fix the problem by surrounding the instruction in a function with
`target_feature` enabled, making sure that function is inlined into the
surrounding GEMM kernel.