Include node name in model load errors where applicable
Make it easier to debug model load errors that relate to specific nodes by
including the name of the node in the error message.

In the process the types of some `ModelLoadError` variants have been erased and
replaced with generic `Box<dyn Error>` types. This enables adding new
information to these errors in future without a breaking API change. This change
is itself technically a breaking change, although it is unlikely anyone was
relying on them.