Optimize Arm GEMM microkernel using FMLA (by element)
Arm supports an FMLA instruction variant [^1] which multiples one lane of a
vector by values in another vector and then accumulates into a destination.
This enables reducing the number of loads from A in each iteration of the GEMM
microkernel by a factor of 4. Instead of broadcasting one scalar value from A at
a time into a register used as an FMA operand, we can load a vector of 4
elements from A and then each FMA specifies which of the 4 elements to use.

The way this is implemented is by using `broadcast_lane` + `mul_add` generic
operations (`vdupq_laneq_f32` followed by `fmla`) which LLVM will fuse into a
single `fmla`.

[^1]: https://developer.arm.com/documentation/ddi0596/2021-03/SIMD-FP-Instructions/FMLA--by-element---Floating-point-fused-Multiply-Add-to-accumulator--by-element--?lang=en