Optimize Arm int8 GEMM using UDOT-by-element
Similar to https://github.com/robertknight/rten/pull/679, use the by-element
variant of UDOT [^1] to reduce the number of loads in the int8 GEMM microkernel.
Instead of performing 4 scalar 32-bit loads and broadcasting the result, perform
one 128-bit load and use indexed UDOT to implicitly broadcast a 32-bit lane for
each of the 4 rows.

[^1]: https://developer.arm.com/documentation/100069/0609/A64-SIMD-Vector-Instructions/UDOT--vector--by-element-