Support fusing subgraphs with multiple outputs
Lift the restriction that fused operators can only have a single output. This
restriction might have simplified the code at an earlier point in time, but that
is no longer the case. None of the current fused operators actually have
multiple outputs, so this doesn't change what the optimizer does.