Remove `x_in_place` functions for most unary operators
As part of addressing https://github.com/robertknight/rten/issues/192, remove
the `*_in_place` variants for unary operators and instead test in-place
operations by calling the `Operator::run_in_place` implementation directly. This
will make it easier to change the behavior of `run_in_place` to fall back to a
non-in-place operation if the input is not contiguous.

The `*_in_place` functions for operators in other modules have also been
un-exported from the `ops` module, to allow for similar changes in future. This
is a breaking change since the `ops` module is public, but there are probably
very few if any downstream consumers of these.