Handle non-contiguous inputs more efficiently in in-place unary ops
When a vectorized and parallelized unary operator's `run_in_place`
implementation is invoked with a non-contiguous input, fall back to using the
non in-place version instead of using a scalar version of the function. This
incurs the overhead of first copying the input into a contiguous buffer, but the
cost is usually much less than forgoing all parallelism. This can be optimized
in future by materializing the contiguous input in cache-sized chunks.

An additional benefit of this change is that the output buffer is nicely
contiguous for subsequent operations.

Fixes https://github.com/robertknight/rten/issues/192