Use a hash map instead of a vec to store graph nodes
This will make it possible for the optimizer to remove unused nodes from the
graph after fusion, or build a new graph with some node IDs omitted.

Using `FxHashMap<NodeId, Node>` resulted in a ~2% slowdown in the GPT2 example
(base model) with more self time spent in `Graph::run_plan`. To mitigate that,
use a custom hasher that uses the u32 node IDs directly as hash values.