Allow constant node IDs to be passed as input IDs for graph runs
Input IDs which refer to constant nodes can be treated the same as value nodes
for the purposes of planning - it is a value which is known to be available when
the plan is executed.

This change will avoid needing to filter out constant nodes when calling into
the planner from the optimizer.